Add : File 0 & 1 of sorting algorithms

+++ 2-selection_sort.c
@@ -11,7 +11,7 @@
  */
 void selection_sort(int *array, size_t size)
 {
-	int i, j, tmp, tmp_idx;
+	size_t i, j, tmp, tmp_idx;
 
 	for (i = 0; i < size; i++)
 	{
